requirements.txt
pandas
flask
numpy
pymongo
python-dotenv
pyyaml
scikit-learn
-e .